Add sqlite db, container type for holdings table, delete button, and validation when adding holdings dialog

+++ holdings-tab.go
@@ -12,35 +12,41 @@ import (
 	"fyne.io/fyne/v2/widget"
 )
 
-func (app *Config)  holdingsTab() *fyne.Container {
+func (app *Config) holdingsTab() *fyne.Container {
+	app.Holdings = app.getHoldingSlice()
 	app.HoldingsTable = app.getHoldingsTable()
 
-	holdingsContainer := container.NewVBox(app.HoldingsTable)
+	holdingsContainer := container.NewBorder(
+		nil,
+		nil,
+		nil,
+		nil,
+		container.NewAdaptiveGrid(1, app.HoldingsTable),
+	)
 
 	return holdingsContainer
 }
 
 func (app *Config) getHoldingsTable() *widget.Table {
-	data := app.getHoldingSlice()
-	app.Holdings = data
-
 	t := widget.NewTable(
 		func() (int, int) {
-			return len(data), len(data[0])
+			return len(app.Holdings), len(app.Holdings[0])
 		},
 		func() fyne.CanvasObject {
 			ctr := container.NewVBox(widget.NewLabel(""))
 			return ctr
 		},
 		func(i widget.TableCellID, o fyne.CanvasObject) {
-			if i.Col == (len(data[0])-1) && i.Row != 0 {
+			if i.Col == (len(app.Holdings[0])-1) && i.Row != 0 {
 				// last cell - put in a button
-				w := widget.NewButtonWithIcon("Delete", theme.DeleteIcon(), func(){
+				w := widget.NewButtonWithIcon("Delete", theme.DeleteIcon(), func() {
 					dialog.ShowConfirm("Delete?", "", func(deleted bool) {
-						id, _ := strconv.Atoi(data[i.Row][0].(string))
-						err := app.DB.DeleteHolding(int64(id))
-						if err != nil {
-							app.ErrorLog.Println(err)
+						if deleted {
+							id, _ := strconv.Atoi(app.Holdings[i.Row][0].(string))
+							err := app.DB.DeleteHolding(int64(id))
+							if err != nil {
+								app.ErrorLog.Println(err)
+							}
 						}
 						// refresh the holdings table
 						app.refreshHoldingsTable()
@@ -54,15 +60,15 @@ func (app *Config) getHoldingsTable() *widget.Table {
 			} else {
 				// we're just putting in textual information
 				o.(*fyne.Container).Objects = []fyne.CanvasObject{
-					widget.NewLabel(data[i.Row][i.Col].(string)),
+					widget.NewLabel(app.Holdings[i.Row][i.Col].(string)),
 				}
 			}
 		})
 
-		colWidths := []float32{50, 200, 200, 200, 110}
-		for i := 0; i < len(colWidths); i++ {
-			t.SetColumnWidth(i, colWidths[i])
-		}
+	colWidths := []float32{50, 200, 200, 200, 110}
+	for i := 0; i < len(colWidths); i++ {
+		t.SetColumnWidth(i, colWidths[i])
+	}
 
 	return t
 }
@@ -77,12 +83,12 @@ func (app *Config) getHoldingSlice() [][]interface{} {
 
 	slice = append(slice, []interface{}{"ID", "Amount", "Price", "Date", "Delete?"})
 
-	for _, x := range holdings{
+	for _, x := range holdings {
 		var currentRow []interface{}
 
 		currentRow = append(currentRow, strconv.FormatInt(x.ID, 10))
 		currentRow = append(currentRow, fmt.Sprintf("%d toz", x.Amount))
-		currentRow = append(currentRow, fmt.Sprintf("$%2f", float32(x.PurchasePrice/100)))
+		currentRow = append(currentRow, fmt.Sprintf("$%.2f", float32(x.PurchasePrice/100)))
 		currentRow = append(currentRow, x.PurchaseDate.Format("2006-01-02"))
 		currentRow = append(currentRow, widget.NewButton("Delete", func() {}))
 
@@ -101,4 +107,3 @@ func (app *Config) currentHoldings() ([]repository.Holdings, error) {
 
 	return holdings, nil
 }
-
